pyproject.toml
[tool.poetry]
name = "avito"
version = "0.1.0"
description = ""
authors = ["y"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.2"
psycopg2 = "^2.9.3"
Pillow = "^9.2.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.1"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.0"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
